+++ tensorflow/cppbuild.sh
@@ -24,7 +24,7 @@ case $PLATFORM in
 esac
 
 PROTOBUF_VERSION=fb714b3606bd663b823f6960a73d052f97283b74
-TENSORFLOW_VERSION=0.8.0rc0
+TENSORFLOW_VERSION=0.8.0
 
 download https://github.com/google/protobuf/archive/$PROTOBUF_VERSION.tar.gz protobuf-$PROTOBUF_VERSION.tar.gz
 download https://github.com/tensorflow/tensorflow/archive/v$TENSORFLOW_VERSION.tar.gz tensorflow-$TENSORFLOW_VERSION.tar.gz
@@ -43,6 +43,6 @@ ln -snf ../../protobuf-$PROTOBUF_VERSION protobuf
 cd ..
 patch -Np1 < ../../../tensorflow-$TENSORFLOW_VERSION.patch
 ./configure
-bazel build -c opt //tensorflow/cc:libtensorflow.so $BUILDFLAGS
+bazel build -c opt //tensorflow/cc:libtensorflow.so $BUILDFLAGS --spawn_strategy=standalone --genrule_strategy=standalone --verbose_failures
 
 cd ../..